+++ 2/two.c
@@ -10,18 +10,18 @@
 
 struct report {
   s32* level;
-  size_t num_levels;
+  u64 num_levels;
 };
 
 struct reports {
   struct report* rep;
-  size_t num_reports;
+  u64 num_reports;
 };
 
 internal void push_level(struct report* report, s32 level) {
   assert(report);
 
-  size_t new_num_levels = report->num_levels + 1;
+  u64 new_num_levels = report->num_levels + 1;
   int* new_level =
       (int*)realloc(report->level, new_num_levels * sizeof(report->level));
 
@@ -34,10 +34,10 @@ internal void push_level(struct report* report, s32 level) {
   report->num_levels = new_num_levels;
 }
 
-internal b8 is_safe(s32* levels, size_t num_levels) {
+internal b8 is_safe(s32* levels, u64 num_levels) {
   b8 increasing = TRUE;
   b8 decreasing = TRUE;
-  for (size_t i = 0; i < num_levels - 1; i++) {
+  for (u64 i = 0; i < num_levels - 1; i++) {
     s32 diff = abs(levels[i] - levels[i + 1]);
     if (diff < 1 || diff > 3) {
       return FALSE;
@@ -60,10 +60,10 @@ internal b8 is_safe(s32* levels, size_t num_levels) {
 }
 
 internal s32* remove_at(int* input_array,
-                        size_t input_array_length,
-                        size_t index_to_remove,
-                        size_t* new_array_length) {
-  size_t new_length = input_array_length - 1;
+                        u64 input_array_length,
+                        u64 index_to_remove,
+                        u64* new_array_length) {
+  u64 new_length = input_array_length - 1;
 
   int* new_array = (int*)malloc(new_length * sizeof(int));
   DJC_MALLOC_CHECK(new_array);
@@ -73,7 +73,7 @@ internal s32* remove_at(int* input_array,
     exit(EXIT_FAILURE);
   }
 
-  for (size_t i = 0, j = 0; i < input_array_length; i++) {
+  for (u64 i = 0, j = 0; i < input_array_length; i++) {
     if (index_to_remove != i) {
       new_array[j++] = input_array[i];
     }
@@ -91,8 +91,8 @@ internal b8 check(struct report* report) {
     return TRUE;
   }
 
-  for (size_t i = 0; i < report->num_levels; i++) {
-    size_t new_array_length = 0;
+  for (u64 i = 0; i < report->num_levels; i++) {
+    u64 new_array_length = 0;
     int* new_array =
         remove_at(report->level, report->num_levels, i, &new_array_length);
     b8 now_safe = is_safe(new_array, new_array_length);
@@ -106,7 +106,7 @@ internal b8 check(struct report* report) {
 
 s32 main(void) {
   struct Arena* mem = arena_create("General", 80);
-  size_t file_size = 0;
+  u64 file_size = 0;
   char* input_file_path = djc_get_input_file("\\resources\\2\\input.txt");
   char* file = djc_load_entire_file(input_file_path, &file_size);
   djc_convert_crlf_to_lf(file);
@@ -121,7 +121,7 @@ s32 main(void) {
     return 0;
   }
 
-  size_t i = 0;
+  u64 i = 0;
   for (char* line = file; line; line = get_next_line(line)) {
     struct report* report = &reports_list.rep[i];
     report->level = mem->current;
@@ -140,7 +140,7 @@ s32 main(void) {
     i++;
   }
 
-  size_t num_safe_reports = 0;
+  u64 num_safe_reports = 0;
 
   for (i = 0; i < reports_list.num_reports; i++) {
     struct report* report = &reports_list.rep[i];
